Fix bug in default comparator: Return correct separator for difficult cases
This is important if TableBuilder wants a separator between K and succ(K), where succ(K) is usually
just a string of K's first characters plus 1. In this case we now return K with its last byte incremented
by one, which isn't really "short" anymore, but a valid separator.
Otherwise, logic relying on the key in index blocks being actually *past* the referenced block fails in
some cases, for example on the last block of a table.